Organized files better
Tip #1:
You don't want to have the main HUD folder inside the main repository folder. You want the main repository folder to be the HUD folder.
So keep files inside `main` instead of making a separate folder and putting all the files inside there. This is because github will make users download a .zip file, which will have the `cherrytf2hudthing-main` inside it, and then `CherryHUD`
The extra folder is unnecessary.